refactor: float->double filters; remove obsolete tv code

+++ src/filterchroma.cpp
@@ -32,7 +32,7 @@ FilterChroma::FilterChroma() {
 FilterChroma::~FilterChroma() {
 }
 
-float FilterChroma::next(const float v) {
+double FilterChroma::next(const double v) {
     xv[0] = xv[1]; xv[1] = xv[2]; xv[2] = xv[3];
     xv[3] = v / GAIN;
     yv[0] = yv[1]; yv[1] = yv[2]; yv[2] = yv[3];
